refactor: split grok tools by responsibility
