+++ packages/agent-core/agent-core.cabal
@@ -49,9 +49,16 @@ library
                     , Agent.Tools.Codex
                     , Agent.Tools.Dangerous
                     , Agent.Tools.Grok
+                    , Agent.Tools.Grok.Common
+                    , Agent.Tools.Grok.Grep
+                    , Agent.Tools.Grok.ListDir
                     , Agent.Tools.Grok.Prompt
+                    , Agent.Tools.Grok.ReadFile
+                    , Agent.Tools.Grok.SearchReplace
                     , Agent.Tools.Grok.Shell
                     , Agent.Tools.Grok.Task
+                    , Agent.Tools.Grok.TaskControl
+                    , Agent.Tools.Grok.Terminal
                     , Agent.Tools.Ghci
                     , Agent.Tools.IO
                     , Agent.Tools.MultiAgents
